% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_mixture_rules.R
\name{average_mixture_rules}
\alias{average_mixture_rules}
\title{Estimate the average rule. This is the rule that is the average
across thresholds in folds and gives the lower and upper bounds}
\usage{
average_mixture_rules(
  group_list,
  data = data,
  mix_comps = mix_comps,
  n_folds,
  mixture_results
)
}
\arguments{
\item{group_list}{List of dataframes grouped by rules for variable sets}

\item{data}{Full data}

\item{mix_comps}{Mixture components of A}

\item{n_folds}{Number of folds used in cross-validation}

\item{mixture_results}{data frame of results found for mixture rules}
}
\value{
Rules object. TODO: add more detail here.

Data frame with mixture results including the union rule and
proportion found across the folds
}
\description{
This function takes in a list of rules that are grouped by
variable sets. These rules for different variable sets may be slightly
different across the folds so we make a average rule for each variable set.
This entails creating a new rule that is is the average of cutpoints:
}
