% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ATE.R
\name{calc_ate_estimates}
\alias{calc_ate_estimates}
\title{Calculate the ATE and variance estimates}
\usage{
calc_ate_estimates(data, ate_var, y, p_adjust_n, v_fold = FALSE)
}
\arguments{
\item{data}{Data frame that includes treatement/exposure}

\item{ate_var}{Variable that the ATE is assigned to in \code{data}}

\item{y}{Variable name in \code{data} for the raw outcome}

\item{p_adjust_n}{Number repeated measures to adjust p-value}

\item{v_fold}{TRUE/FALSE whether to calculate v-fold specific ATE estimates}
}
\value{
A \code{numeric} vector of the same length as \code{vals}, where
the returned values are bounded to machine precision. This is intended to
avoid numerical instability issues.
}
\description{
Calculate the ATE and variance estimates
}
\details{
Calculates the average treatment effect and variance estimates
from the influence curve across folds
}
