% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_marginal_rule_rmses}
\alias{calc_marginal_rule_rmses}
\title{Calculate the mean RMSE in each marginal group}
\usage{
calc_marginal_rule_rmses(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Calculate the mean RMSE in each marginal group
}
