% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_mixture_rule_RMSEs}
\alias{calc_mixture_rule_RMSEs}
\title{Calculate the mean RMSE in each interaction group}
\usage{
calc_mixture_rule_RMSEs(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Calculate the mean RMSE in each interaction group
}
