% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mixtures_nde.R
\name{calc_mixtures_nde}
\alias{calc_mixtures_nde}
\title{Calculate the NDE for each mixture rule}
\usage{
calc_mixtures_nde(
  input_mix_rules,
  input_mix_data,
  y,
  n_folds,
  no_mixture_rules
)
}
\arguments{
\item{input_mix_rules}{List of dataframes of rules found for a mixture
across the folds}

\item{input_mix_data}{Nuisance parameter data for mixture rules found}

\item{y}{Outcome variable
across the folds}

\item{n_folds}{Number of folds used in cross-validation}

\item{no_mixture_rules}{TRUE/FALSE whether no mixture rules were found
across all the folds}
}
\value{
A list with mixture analysis results which includes:
\itemize{
\item \code{results}: A data frame with variable threshold combinations
on the rows and ATE, variance and consistency estimates on the columns.
\item \code{group_list}: A list of rule combinations found in the
ensemble decision tree model grouped by variable sets in the rules and
directions of the coefficient in the linear model. Also provided is the
fold the rule was found and the RMSE of the model which used the rule.
\item \code{mixture_data_list}: A list of data frames which houses the
data for each mixture rule evaluated as an exposure, the baseline
covariates, outcome, and nuisance parameter estimates for the respective
rule.
}
}
\description{
Aggregate mixture rules found across the folds that have
the same variables. For each rule
extract the relevant nuisance parameter data calculated in the folds.
Given the validation data estimates across
the folds, for each tree do a TMLE update step to target the natural
direct effect. Update the initial counterfactuals,
calculate the influence curve and using the influence curve calculate
variance estimates and p-values.
}
