% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_v_fold_mixture_nie.R
\name{calc_v_fold_mixtures_nie}
\alias{calc_v_fold_mixtures_nie}
\title{Calculate the NIE for V-fold specific rules}
\usage{
calc_v_fold_mixtures_nie(input_mix_rules, input_mix_data, y, n_folds)
}
\arguments{
\item{input_mix_rules}{List of dataframes of rules found for a mixture
across the folds}

\item{input_mix_data}{Nuisance parameter data for mixture rules found
across the folds}

\item{y}{Character indicating the outcome variable}

\item{n_folds}{Number of folds used in cross-validation}
}
\value{
Rules object. TODO: add more detail here.

A data frame of mixture results for each variable level combination
for each fold
}
\description{
For each rule in each fold extract the relevant nuisance
parameter data calculated in the folds. Given the validation data estimates
for each rule do a TMLE update step to target the natural indirect effect.
Update the initial counterfactuals,calculate the influence curve
and using the influence curve calculate variance estimates and p-values.
}
