% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculatePooledEstimate}
\alias{calculatePooledEstimate}
\title{Calculates the Inverse Variance Pooled Estimate Including Null Folds}
\usage{
calculatePooledEstimate(est, se, n_folds, n_intxn)
}
\arguments{
\item{n_folds}{Total number of folds specified}

\item{results_df}{Table of results}
}
\description{
Does a weighted combination estimate for folds with estimates
and null folds finding 0 using inverse variance
}
