% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_sticker_v2.R
\docType{data}
\name{col_bg}
\alias{col_bg}
\title{Creates the CVtreeMLE sticker from the image in figures}
\format{
An object of class \code{character} of length 1.
}
\usage{
col_bg
}
\description{
Creates the CVtreeMLE sticker from the image in figures
}
\keyword{datasets}
