% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_common_mixture_rule.R
\name{common_mixture_rules}
\alias{common_mixture_rules}
\title{Estimate the union rule. This is the rule that covers all observations
across the folds.}
\usage{
common_mixture_rules(
  group_list,
  data = data,
  mix_comps = mix_comps,
  mixture_results,
  n_folds,
  no_mixture_rules
)
}
\arguments{
\item{group_list}{List of dataframes grouped by rules for variable sets}

\item{data}{Full data}

\item{mix_comps}{Mixture components of A}

\item{mixture_results}{data frame of results found for mixture rules}

\item{n_folds}{Number of folds used in cross-validation}

\item{no_mixture_rules}{TRUE/FALSE if no mixture rule was found}
}
\value{
Rules object. TODO: add more detail here.

Data frame with mixture results including the union rule and
proportion found across the folds
}
\description{
This function takes in a list of rules that are grouped by
variable sets. These rules for different variable sets may be slightly
different across the folds so we make a union rule for each variable set.
This entails creating a new rule that is essentially:
(rule fold 1) OR (rule fold 2) OR (rule fold 3) etc. We then evaluate this
rule on the input data to create a binary indicator of the union rule.
Then, for variables used in the rule, find the min in and max in regions
indicated by the rule for each variable. We then put these together with
AND statements to create the union rule.
}
