% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_meta_mix_results.R
\name{compute_meta_mix_results}
\alias{compute_meta_mix_results}
\title{Compute v-fold specific estimates and do a meta-analysis type pooled average ATE for mixture rules.}
\usage{
compute_meta_mix_results(v_fold_mixture_results, mix_comps, n_folds, data)
}
\arguments{
\item{v_fold_mixture_results}{List of dataframes for v-fold specific estimates of the fold-specific rule results}

\item{mix_comps}{Vector of characters indicating the mixture components}

\item{n_folds}{Number of folds used in cross-validation}

\item{data}{Input dataframe on which to evaluate the rules}
}
\value{
Rules object. TODO: add more detail here.
}
\description{
For each fold, estimates the ATE for a fold specific mixture rule. Also estimates a meta-analysis type average ATE and pooled variance. Creates a union rule that covers all the folds in the rules.
}
