% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_rules}
\alias{create_rules}
\title{From HAL fit, create string based rules}
\usage{
create_rules(basis_list, col_names)
}
\arguments{
\item{basis_list}{Basis list from HAL}

\item{col_names}{Column names}
}
\description{
From HAL fit, create string based rules
}
\details{
Inputs the basis list from HAL then returns a list of rules
}
