% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVtreeMLE.R
\name{CVtreeMLE}
\alias{CVtreeMLE}
\title{Estimate the joint impact of a mixed exposure using cross-validated targeted learning and decision trees}
\usage{
CVtreeMLE(
  W,
  A,
  Y,
  data,
  SL.library,
  n_folds,
  family,
  H.AW_trunc_lvl,
  n_boot_ensemble,
  n_stable_tree,
  minbucket,
  verbose,
  parallel
)
}
\arguments{
\item{W}{A vector of characters indicating which variables in the data to use as covariates.}

\item{A}{A vector of characters indicating which variables in the data to use as exposures.}

\item{Y}{A character indicating which variable in the data to use as the outcome.}

\item{data}{Data frame of (W,A,Y) variables of interest.}

\item{SL.library}{Library of algorithms used by Super Learner for estimating both the propensity of exposure and the outcome.}

\item{n_folds}{Number of cross-validation folds.}

\item{family}{Family ('binomial' or 'gaussian').}

\item{H.AW_trunc_lvl}{Truncation level for the clever covariate.}

\item{n_boot_ensemble}{Number of bootstrap fits used to identify consistent partitions in the mixture space.}

\item{n_stable_tree}{Number of bootstraps a set of variables must be in to be considered a consistent tree for the mixture.}

\item{minbucket}{Minimum number of observations needed in a bin to make a partition in the marginal decision trees.}

\item{verbose}{If true, creates a sink file where detailed information and diagnostics of the run process are given.}

\item{parallel}{Use parallel processing if a backend is registered; enabled by default.}
}
\value{
Results object. TODO: add more detail here.
}
\description{
Estimate the joint impact of a mixed exposure using cross-validated targeted learning and decision trees
}
\section{Authors}{

David McCoy, University of California, Berkeley
}

\section{References}{

Benjamini, Y., & Hochberg, Y. (1995). \emph{Controlling the false discovery
rate: a practical and powerful approach to multiple testing}. Journal of the
royal statistical society. Series B (Methodological), 289-300.

Gruber, S., & van der Laan, M. J. (2012). \emph{tmle: An R Package for
Targeted Maximum Likelihood Estimation}. Journal of Statistical Software,
51(i13).

Hubbard, A. E., Kherad-Pajouh, S., & van der Laan, M. J. (2016).
\emph{Statistical Inference for Data Adaptive Target Parameters}. The
international journal of biostatistics, 12(1), 3-19.

Hubbard, A., Munoz, I. D., Decker, A., Holcomb, J. B., Schreiber, M. A.,
Bulger, E. M., ... & Rahbar, M. H. (2013). \emph{Time-Dependent Prediction
and Evaluation of Variable Importance Using SuperLearning in High Dimensional
Clinical Data}. The journal of trauma and acute care surgery, 75(1 0 1), S53.

Hubbard, A. E., & van der Laan, M. J. (2016). \emph{Mining with inference:
data-adaptive target parameters (pp. 439-452)}. In P. Buhlmann et al. (Ed.),
\emph{Handbook of Big Data}. CRC Press, Taylor & Francis Group, LLC: Boca
Raton, FL.

van der Laan, M. J. (2006). \emph{Statistical inference for variable
importance}. The International Journal of Biostatistics, 2(1).

van der Laan, M. J., & Pollard, K. S. (2003). \emph{A new algorithm for
hybrid hierarchical clustering with visualization and the bootstrap}. Journal
of Statistical Planning and Inference, 117(2), 275-303.

van der Laan, M. J., Polley, E. C., & Hubbard, A. E. (2007). \emph{Super
learner}. Statistical applications in genetics and molecular biology, 6(1).

van der Laan, M. J., & Rose, S. (2011). \emph{Targeted learning: causal
inference for observational and experimental data}. Springer Science &
Business Media.
}

