% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_cum_sum_exposure.R
\name{est_cum_sum_exposure}
\alias{est_cum_sum_exposure}
\title{Calculate the cumulative sum exposure based on marginals found in the fold}
\usage{
est_cum_sum_exposure(At_c, Av_c, W, SL.library, H.AW_trunc_lvl)
}
\arguments{
\item{At_c}{Training data}

\item{Av_c}{Validation data}

\item{W}{Vector of characters denoting covariates}

\item{SL.library}{Super Learner library for fitting Q (outcome mechanism) and g (treatment mechanism)}

\item{H.AW_trunc_lvl}{Truncation level of the clever covariate (induces more bias to reduce variance)}
}
\value{
Rules object. TODO: add more detail here.
}
\description{
For each marginal mixture component rule found, create a g estimator for the probability of being exposed to the rule thresholds,
and a Q estimator for the outcome E(Y| A = a_mix, W). Get estimates of g and Q using the validation data and
calculate the clever covariate used in the TMLE fluctuation step.
}
