% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_nuisance_parameters.R
\name{est_mix_nuisance_params}
\alias{est_mix_nuisance_params}
\title{For each mixture rule found, create a g estimator for the probability of being exposed to the rule thresholds,
and a Q estimator for the outcome E(Y| A = a_mix, W). Get estimates of g and Q using the validation data and
calculate the clever covariate used in the TMLE fluctuation step.}
\usage{
est_mix_nuisance_params(
  At,
  Av,
  W,
  no_mix_rules,
  Q1_stack,
  family,
  rules,
  H.AW_trunc_lvl
)
}
\arguments{
\item{At}{Training data}

\item{Av}{Validation data}

\item{W}{Vector of characters denoting covariates}

\item{no_mix_rules}{TRUE/FALSE indicator for if no mixture rules were found}

\item{Q1_stack}{Super Learner library for fitting Q (outcome mechanism) and g (treatment mechanism)}

\item{family}{Binomial or gaussian}

\item{rules}{Dataframe of rules found during the PRE fitting process}

\item{H.AW_trunc_lvl}{Truncation level of the clever covariate (induces more bias to reduce variance)}
}
\value{
Rules object. TODO: add more detail here.
}
\description{
For each mixture rule found, create a g estimator for the probability of being exposed to the rule thresholds,
and a Q estimator for the outcome E(Y| A = a_mix, W). Get estimates of g and Q using the validation data and
calculate the clever covariate used in the TMLE fluctuation step.
}
