% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_marginal_rules.R
\name{evaluate_marginal_rules}
\alias{evaluate_marginal_rules}
\title{Evaluate mixture rules found during the rpart decision tree process}
\usage{
evaluate_marginal_rules(data, marg_decisions, no_marg_rules, mix_comps)
}
\arguments{
\item{data}{Input data}

\item{marg_decisions}{List of rules determined for each mixture component}

\item{no_marg_rules}{TRUE/FALSE if no marginal rules were found across the
folds}

\item{mix_comps}{Vector of characters denoting mixture components}
}
\value{
A list with marginal rules evaluated which includes:
\itemize{
\item \code{data}: The fold raw data with the sum of evaluated rule
binary vectors for each mixture component added.
\item \code{marg_rule_df}: A binary matrix where each marginal rule has
been evaluated to a binary vector.
}
}
\description{
Evaluate mixture rules found during the rpart decision tree process
}
