% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_mix_rules.R
\name{evaluate_mixture_rules}
\alias{evaluate_mixture_rules}
\title{Evaluate mixture rules found during the PRE process}
\usage{
evaluate_mixture_rules(data, rules)
}
\arguments{
\item{data}{Input data}

\item{rules}{Dataframe of the rules determined}
}
\value{
Rules object. TODO: add more detail here.
}
\description{
Evaluate mixture rules found during the PRE process
}
