% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate_rules_to_binary}
\alias{evaluate_rules_to_binary}
\title{Evaluate Rules to Binary Indicators}
\usage{
evaluate_rules_to_binary(rules, data, Y)
}
\arguments{
\item{rules}{List of rules}

\item{data}{Data to evaluate rules on}

\item{Y}{Outcome that is appended to the binary data}
}
\description{
Evaluate Rules to Binary Indicators
}
\details{
Takes in a list of rules and outputs a binary matrix
}
