% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_marginal_rules}
\alias{filter_marginal_rules}
\title{Filter marginal rules across the folds for only those that have the same variables}
\usage{
filter_marginal_rules(data, n_folds)
}
\arguments{
\item{data}{Input data}

\item{n_folds}{Number of folds in the CV}
}
\description{
Filter marginal rules across the folds for only those that have the same variables
}
