% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_mixture_rules}
\alias{filter_mixture_rules}
\title{Filter mixture rules across the folds for only those that have the same variables and directions across all folds}
\usage{
filter_mixture_rules(data, n_folds)
}
\arguments{
\item{data}{Input data}

\item{n_folds}{Number of folds in the CV}
}
\description{
Filter mixture rules across the folds for only those that have the same variables and directions across all folds
}
