% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_rules}
\alias{filter_rules}
\title{Filter data based on fold}
\usage{
filter_rules(data, fold_k)
}
\arguments{
\item{data}{Input data}

\item{fold_k}{Current fold}
}
\description{
Filter data based on fold
}
\details{
Filter data to only the training fold of interest during CV
}
