% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_iterative_mix_rule_backfitting.R
\name{fit_iterative_mix_rule_backfitting}
\alias{fit_iterative_mix_rule_backfitting}
\title{Iteratively Backfit a Super Learner, h(x) = Y|W, and an Ensemble Decision Algorithm, g(x), Y|M_x until Convergence.}
\usage{
fit_iterative_mix_rule_backfitting(At, A, W, Y, Q1_stack, fold)
}
\arguments{
\item{At}{Training dataframe}

\item{A}{Variable names in the mixture}

\item{W}{Variable names in the covariates}

\item{Y}{Variable name for the outcome}

\item{Q1_stack}{Stack of algorithms made in SL 3 used in ensemble machine learning to fit Y|W}

\item{fold}{Current fold in the cross-validation}
}
\value{
Rules object. TODO: add more detail here.
}
\description{
Iteratively Backfit a Super Learner, h(x) = Y|W, and an Ensemble Decision Algorithm, g(x), Y|M_x until Convergence.
}
\details{
Performs an iterative backfitting algorithm to flexibly adjust for covariates W while finding the best fitting set of mixture rules to partition the space in M.
}
