% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ATE.R
\name{fit_least_fav_submodel}
\alias{fit_least_fav_submodel}
\title{Least Favorable Submodel}
\usage{
fit_least_fav_submodel(H.AW, data, QbarAW, Qbar1W, Qbar0W)
}
\arguments{
\item{H.AW}{\code{numeric} vector of values for the clever covariate used in the fluctuation model}

\item{data}{dataframe that includes treatement/exposure}

\item{QbarAW}{Initial predictions for Y|A,W - A being observed A}

\item{Qbar1W}{Initial predictions for Y|A = 1,W - A being deterministically set to 1}

\item{Qbar0W}{Initial predictions for Y|A = 1,W - A being deterministically set to 0}
}
\value{
A \code{numeric} vector of the same length as \code{vals}, where
the returned values are bounded to machine precision. This is intended to
avoid numerical instability issues.
}
\description{
Least Favorable Submodel
}
\details{
Does flucutation of initial outcome estimates using a least favorable submodel in the TMLE update step
}
