% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_decision_trees.R
\name{fit_marginal_decision_trees}
\alias{fit_marginal_decision_trees}
\title{Fit a decision tree on the remaining variance of Y after removing variance attributable to W
and other non-target mixture components M}
\usage{
fit_marginal_decision_trees(
  mix_comps,
  At,
  covariates,
  SL.library,
  family,
  minbucket
)
}
\arguments{
\item{mix_comps}{A vector of characters indicating variables for the mixture components}

\item{At}{Training data}

\item{covariates}{A vector of characters indicating variables that are covariates
on the mixture data to predict remaining variance in Y}

\item{SL.library}{The Super Learner library to be used to remove the variance in Y|A,W}

\item{family}{Family ('binomial' or 'gaussian')}

\item{minbucket}{The minimum number of observations in any terminal <leaf> node.}
}
\value{
Rules object. TODO: add more detail here.
}
\description{
Currently uses rpart package but could be generalized
}
