% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_iterative_mix_rule_backfitting.R
\name{fit_mix_rule_backfitting}
\alias{fit_mix_rule_backfitting}
\title{Iteratively Backfit a Super Learner, h(x) = Y|W, and an Ensemble
Decision Algorithm, g(x), Y|M_x until Convergence.}
\usage{
fit_mix_rule_backfitting(
  at,
  a,
  w,
  y,
  w_stack,
  fold,
  max_iter,
  verbose,
  parallel_cv,
  seed
)
}
\arguments{
\item{at}{Training dataframe}

\item{a}{Variable names in the mixture}

\item{w}{Variable names in the covariates}

\item{y}{Variable name for the outcome}

\item{w_stack}{Stack of algorithms made in SL 3 used in ensemble machine
learning to fit Y|W}

\item{fold}{Current fold in the cross-validation}

\item{max_iter}{Max number of iterations of iterative backfitting algorithm}

\item{verbose}{Run in verbose setting}

\item{parallel_cv}{TRUE/FALSE indicator to parallelize cv}

\item{seed}{Seed number for consistent results}
}
\value{
Rules object. TODO: add more detail here.
}
\description{
Iteratively Backfit a Super Learner, h(x) = Y|W, and an Ensemble
Decision Algorithm, g(x), Y|M_x until Convergence.
}
\details{
Performs an iterative backfitting algorithm to flexibly adjust
for covariates W while finding the best fitting set of mixture rules
to partition the space in M.
}
\examples{
data <- simulate_mixture_cube()
data$y_scaled <- data$y
mix_comps <- c("M1", "M2", "M3")
W <- c("w", "w2")
sls <- create_sls()
w_stack <- sls$W_stack
tree_stack <- sls$A_stack
example_output <- fit_mix_rule_backfitting(at = data,
                                            a = mix_comps,
                                            w = W,
                                            y = "y_scaled",
                                            w_stack = w_stack,
                                            fold = 1,
                                            max_iter = 1,
                                            verbose = FALSE,
                                            parallel = FALSE,
                                            seed = 6442)
}
