% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mix_rule_ensemble.R
\name{fit_mix_rule_ensemble}
\alias{fit_mix_rule_ensemble}
\title{Fit decision tree ensembles on the remaining variance of Y using all mixture components as features}
\usage{
fit_mix_rule_ensemble(At, formula, rule_boot_n, n_stable_trees, A)
}
\arguments{
\item{At}{Training dataframe}

\item{formula}{Formula passed to the pre tree fitting algorithm}

\item{rule_boot_n}{Number of times to resample data with replacement and refit the pre algorithm
on the mixture data to predict remaining variance in Y}

\item{n_stable_trees}{Number of times a set of variables must appear in the same rule across the bootstrap to be considered stable
and move to downstream analysis}

\item{A}{Variable names in the mixture}
}
\value{
Rules object. TODO: add more detail here.
}
\description{
Currently uses PRE Predictive Rule Ensembles package but could be generalized
}
