% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_post_counterfactuals.R
\name{fit_post_counterfactuals}
\alias{fit_post_counterfactuals}
\title{Given a vector of target mixture components, load a \code{CVtreeMLE} results object, get the lists of rules, nuisance parameters, and learners used. Apply counterfactuals to the target
variables of interest and return a dataframe with the ATE, variance estimates and p-values for these counterfactuals.}
\usage{
fit_post_counterfactuals(
  modeling_results,
  target_mixtures,
  H.AW_trunc_lvl,
  SL.library,
  p_adjust_n
)
}
\arguments{
\item{modeling_results}{A \code{CVtreeMLE} results object}

\item{target_mixtures}{Vector of characters indicating which mixture compnents to calculate a new counterfactual parameter on given rules found across the folds for each marginal
mixture component}

\item{H.AW_trunc_lvl}{Truncation level of the clever covariate - reduces variance but increases bias of the ATE}

\item{SL.library}{Library of algorithms used to fit the counterfactuals}

\item{p_adjust_n}{p-value adjustment for multiple comparisons}
}
\value{
Rules object. TODO: add more detail here.
}
\description{
Currently uses PRE Predictive Rule Ensembles package but could be generalized
}
