% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flip_rule}
\alias{flip_rule}
\title{Flip Negative Mixture Rules}
\usage{
flip_rule(rules, at, mix_comps)
}
\arguments{
\item{rules}{Row of absolute max rule results from pre}

\item{at}{training data to evaluate rule on}

\item{mix_comps}{mixture components}
}
\description{
Flip Negative Mixture Rules
}
