% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_RMSE_for_models.R
\name{format_RMSE_for_models}
\alias{format_RMSE_for_models}
\title{Creates the RMSE output}
\usage{
format_RMSE_for_models(
  marginal_results,
  mixture_results,
  additive_RMSE_star,
  marg_combo_data
)
}
\arguments{
\item{marginal_results}{Marginal results dataframe}

\item{mixture_results}{Mixture results dataframe}

\item{additive_RMSE_star}{Additive marginal model RMSE}

\item{marg_combo_data}{Non-additive marginal dataframe}
}
\description{
Create a simple data frame with the marginal, mixture, additive marginal and non-additive marginal models and their root mean squared error.
}
