% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_impute.R
\name{impute_NA_vals}
\alias{impute_NA_vals}
\title{Impute NA values and create indicator variables}
\usage{
impute_NA_vals(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Impute NA values and create indicator variables
}
