% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list.rules.party}
\alias{list.rules.party}
\title{Get rules from partykit object in rule fitting}
\usage{
list.rules.party(x, i = NULL, ...)
}
\arguments{
\item{x}{Partkit glmtree model object}

\item{i}{null}

\item{...}{additional arguments}
}
\value{
Vector of rules
}
\description{
Get rules from partykit object in rule fitting
}
