% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{marginal_group_split}
\alias{marginal_group_split}
\title{Group split by marginal variable}
\usage{
marginal_group_split(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Group split by marginal variable
}
