% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHANES_eurocim_data.R
\docType{data}
\name{NHANES_eurocim}
\alias{NHANES_eurocim}
\title{NHANES 2001-2002, POP Exposure on Telomere Length}
\format{
\subsection{\code{NHANES_eurocim}}{

A data frame with 1,330 rows and 33 columns:
}
}
\usage{
NHANES_eurocim
}
\description{
A subset of data from NHANES to reproduce analysis used by Mitro et. al.,
and Gibson et. al.
}
\keyword{datasets}
