% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mixture_results.R
\name{plot_mixture_results}
\alias{plot_mixture_results}
\title{Create dot-whisker plots for the mixture results found}
\usage{
plot_mixture_results(v_intxn_results)
}
\arguments{
\item{v_intxn_results}{Table of the interaction results for a mixture}
}
\description{
Create dot-whisker plots for the mixture results found
}
