% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_out_rule_vars}
\alias{pull_out_rule_vars}
\title{Pull rules out of results table of pre results}
\usage{
pull_out_rule_vars(x, a)
}
\arguments{
\item{x}{Row of a dataframe that contains output from the pre fit}

\item{a}{Vector of characters indicating column names for the mixture
variables}
}
\value{
Vector of matches of mixture variables
}
\description{
Pull rules out of results table of pre results
}
\details{
Simply function that looks at variables used in the pre model
for matches
with the mixture variables and pulls them out to check for consistent
sets of rules found over the iterative backfitting procedure
}
