% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mixture_cube.R
\name{simulate_mixture_cube}
\alias{simulate_mixture_cube}
\title{Simulate a mixture cube to test \code{CVtreeMLE} against simulated ground-truth.}
\usage{
simulate_mixture_cube(
  n_obs,
  splits,
  mins,
  maxs,
  mu,
  sigma,
  w1_betas,
  w2_betas,
  mix_subspace_betas,
  subspace_assoc_strength_betas,
  marginal_impact_betas,
  eps_sd,
  binary
)
}
\arguments{
\item{n_obs}{Number of observations for which to generate data}

\item{splits}{Vector indicating where thresholds should be placed for each mixture variable}

\item{mins}{Vector indicating the minimum values for each mixture variable}

\item{maxs}{Vector indicating the maximum values for each mixture variable}

\item{mu}{Vector indicating the mean values for each mixture variable}

\item{sigma}{Matrix of the variance-covariance structure used to generate the mixture variables}

\item{w1_betas}{Vector of betas that define the subspace probability relationship with covariate W1}

\item{w2_betas}{Vector of betas that define the subspace probability relationship with covariate W2}

\item{mix_subspace_betas}{Vector of betas that define the subspace probabilities}

\item{subspace_assoc_strength_betas}{The outcome Y generated by each partition of the mixture cube}

\item{marginal_impact_betas}{Vector of betas that define the marginal impact each mixture variable has}

\item{eps_sd}{Random error included in the generation of Y}

\item{binary}{TRUE/FALSE depending on if the outcome should be binary}
}
\value{
Rules object. TODO: add more detail here.
}
\description{
Simulate a mixture cube. This creates three correlated mixture variables that are associated with two confounders W1 and W2. First mixtures are generated from a multivariate normal.
A multinomial outcome is generated based on betas input for W1 and W2 - associating each W with a part of the mixture cube. In each part of the mixture cube, transform the multivariate
normal mixture to a uniform distribution, respecting the bounds for parts of the cube. A three variable cube with one threshold per variable has 8 subspaces. An outcome is then generated
as a linear combination of different subspaces.
}
