% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{v_fold_marginal_qgroup_split}
\alias{v_fold_marginal_qgroup_split}
\title{v-fold marginal group split}
\usage{
v_fold_marginal_qgroup_split(data)
}
\arguments{
\item{data}{Input data}
}
\description{
v-fold marginal group split
}
