% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mix_nuisance_params.R
\name{est_mix_nuisance_params}
\alias{est_mix_nuisance_params}
\title{Estimate nuisance parameters for each mixture interaction identified}
\usage{
est_mix_nuisance_params(
  at,
  av,
  w,
  a,
  y,
  no_mix_rules,
  aw_stack,
  family,
  rules,
  parallel_cv,
  seed,
  h_aw_trunc_lvl
)
}
\arguments{
\item{at}{Training data}

\item{av}{Validation data}

\item{w}{Vector of characters denoting covariates}

\item{y}{The outcome variable}

\item{no_mix_rules}{TRUE/FALSE indicator for if no mixture rules were found}

\item{aw_stack}{Super Learner library for fitting Q (outcome mechanism) and
g (treatment mechanism)}

\item{family}{Binomial or continuous}

\item{rules}{Dataframe of rules found during the PRE fitting process}

\item{parallel_cv}{TRUE/FALSE if cv parallelization is used}

\item{seed}{Seed number}

\item{h_aw_trunc_lvl}{Truncation level of the clever covariate (induces more
bias to reduce variance)}
}
\value{
A list of dataframes where the nuisance parameters are added to
the raw data.
}
\description{
For each mixture mixture interaction found, create a g estimator
for the probability of being exposed to the rule thresholds,
and a Q estimator for the outcome E(Y| A = a_mix, W). Get estimates of g
and Q using the validation data and
calculate the clever covariate used in the TMLE fluctuation step.
}
