% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ATE.R
\name{fit_least_fav_submodel}
\alias{fit_least_fav_submodel}
\title{Least Favorable Submodel}
\usage{
fit_least_fav_submodel(h_aw, data, y, qbar_aw, qbar_1w, qbar_0w)
}
\arguments{
\item{h_aw}{\code{numeric} vector of values for the clever covariate}

\item{data}{dataframe that includes treatement/exposure}

\item{y}{Outcome variable
used in the fluctuation model}

\item{qbar_aw}{Initial predictions for Y|A,W - A being observed A}

\item{qbar_1w}{Initial predictions for Y|A = 1,W - A being deterministically
set to 1}

\item{qbar_0w}{Initial predictions for Y|A = 1,W - A being deterministically
set to 0}
}
\value{
A \code{numeric} vector of the same length as \code{vals}, where
the returned values are bounded to machine precision. This is intended to
avoid numerical instability issues.
}
\description{
Least Favorable Submodel
}
\details{
Does flucutation of initial outcome estimates using a least
favorable submodel in the TMLE update step
}
