% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_min_ave_tree.R
\name{fit_min_ave_tree_algorithm}
\alias{fit_min_ave_tree_algorithm}
\title{Fit minimum average tree}
\usage{
fit_min_ave_tree_algorithm(
  at,
  a,
  w,
  y,
  family,
  fold,
  parallel_cv,
  min_max,
  min_obs
)
}
\arguments{
\item{at}{Training dataframe}

\item{a}{Variable names in the mixture}

\item{w}{Variable names in the covariates}

\item{y}{Variable name for the outcome}

\item{fold}{Current fold in the cross-validation}

\item{parallel_cv}{TRUE/FALSE indicator to parallelize cv}

\item{direction}{Positive/negative - max or min coefficient to keep in
the ensemble}

\item{w_stack}{Stack of algorithms made in SL 3 used in ensemble machine
learning to fit Y|W}

\item{max_iter}{Max number of iterations of iterative backfitting algorithm}

\item{verbose}{Run in verbose setting}

\item{seed}{Seed number for consistent results}
}
\value{
A list of the mixture rule results within a fold including:
\itemize{
\item \code{rules}: A data frame with the data adpatively
determined rules found in the \code{pre} model along with the coefficient,
direction, fold, RMSE and other measures.
\item \code{model}: The best fitting pre model found in the fold.
}
}
\description{
Fit minimum average tree
}
\details{
Fits the min ave tree
}
\examples{
data <- simulate_mixture_cube()
mix_comps <- c("M1", "M2", "M3")
W <- c("age", "sex", "bmi")
sls <- create_sls()
w_stack <- sls$W_stack
tree_stack <- sls$A_stack
example_output <- fit_pre_algorithm(
  at = data,
  a = mix_comps,
  w = W,
  y = "y",
  direction = "positive",
  w_stack = w_stack,
  fold = 1,
  max_iter = 1,
  verbose = FALSE,
  parallel = FALSE,
  seed = 6442
)
}
