% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_impute.R
\name{impute_NA_vals}
\alias{impute_NA_vals}
\title{Impute Missing Values And Create Indicator Variables}
\usage{
impute_NA_vals(data, W, A, Y)
}
\arguments{
\item{data}{Input Data passed into \code{CVtreeMLE}}

\item{W}{Vector of Covariates}

\item{A}{Vector of Exposures}

\item{Y}{Outcome}
}
\value{
Data With Missing Values Imputed At The Mean And Indicator Variables for Imputed Columns Added To \code{W}
}
\description{
Impute Missing Values And Create Indicator Variables
}
